requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.0.1
breadability==0.1.20
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.1.0
django-redis==5.4.0
django-rest-auth==0.9.5
django-tenant-schemas==1.12.0
django-tenants==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
exceptiongroup==1.1.2
idna==3.4
iniconfig==2.0.0
joblib==1.3.2
lxml==4.9.3
newsdataapi==0.1.8
nltk==3.8.1
ordered-set==4.1.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.9
pycountry==22.3.5
PyJWT==1.7.1
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
redis==5.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sumy==0.11.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
